Finish Beit Yitzhak web project

+++ web/part3/announcements.php
@@ -1,13 +1,17 @@
 <?php
-$result = $mysqli->query("SELECT message.*,user.first_name FROM message JOIN user ON message.user_id = user.id ORDER BY message.date_created DESC LIMIT 5") or die($mysqli->error);
+$result = $mysqli->query("SELECT message.*,user.first_name FROM message JOIN user ON message.user_id = user.id ORDER BY message.date_created DESC") or die($mysqli->error);
+$messages = array();
+while ($message = $result->fetch_array()) {
+    $messages[] = $message;
+}
 ?>
 <section id="announcements">
-  <a href="javascript:void(0);"><h2>הודעות</h2></a>
+  <h2>הודעות</h2>
   <ul>
-    <?php while ($message = $result->fetch_array()) { ?>
-    <li>
-        <?php if (isset($_SESSION["user"]) && $_SESSION["user"]["id"] == $message["user_id"]) { ?>
-        <span class="close"><a href="index.php?page=<?php echo $page;?>&delete-announcement&id=<?php echo $message["id"];?>">מחק</a></span>
+    <?php foreach ($messages as $idx => $message) { ?>
+    <li class="message <?php echo ($idx >= 3) ? "hidden" : "unhidden";?>">
+        <?php if (isset($_SESSION["user"]) && ($_SESSION["user"]["id"] == $message["user_id"] || $_SESSION["user"]["type"] == 2)) { ?>
+        <span class="close"><a onclick="return confirm('האם אתה בטוח שברצונך למחוק את ההודעה הזו?');" href="index.php?page=<?php echo $page;?>&delete-announcement&id=<?php echo $message["id"];?>">מחק</a></span>
         <?php } ?>
         <span class="title"><?php echo $message["title"];?></span>
         <p>
@@ -16,6 +20,7 @@ $result = $mysqli->query("SELECT message.*,user.first_name FROM message JOIN use
         </p>
     </li>
     <?php } ?>
+    <?php if (isset($_SESSION["user"]) && $_SESSION["user"]["type"] != 0) { ?>
     <li>
         <a href="javascript:void(0);" class="post_message">הוסף הודעה</a>
         <form id="post_message" action="index.php?page=<?php echo $page;?>" method="POST" style="display: none;">
@@ -26,5 +31,6 @@ $result = $mysqli->query("SELECT message.*,user.first_name FROM message JOIN use
             </ul>
         </form>
     </li>
+    <?php } ?>
   </ul>
 </section>
